Move validator to tests
