+++ test/Microsoft.Extensions.Options.Test/Validator/OptionsValidatorOptions.cs
